マップエディタWebComponentを別パッケージに切り出す
