+++ packages/map-editor-components/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "@piyoppi/pico2map-tiled",
+  "name": "@piyoppi/pico2map-ui-components",
   "version": "0.0.17",
-  "description": "",
+  "description": "pico2map Web Components",
   "main": "dist/main.js",
   "repository": {
     "type": "git",
     "url":"https://github.com/piyoppi/pico2map"
   },
   "dependencies": {
+    "lit-element": "^2.4.0",
     "typescript": "^4.1.3"
   },
   "devDependencies": {
@@ -16,7 +17,9 @@
     "ts-jest": "^26.4.4",
     "ts-loader": "^8.0.14",
     "webpack": "^5.17.0",
-    "webpack-cli": "^4.4.0"
+    "webpack-cli": "^4.4.0",
+    "@piyoppi/pico2map-tiled": "0.0.17",
+    "@piyoppi/pico2map-editor": "0.0.17"
   },
   "scripts": {
     "test": "jest",
