+++ packages/map-editor-components/src/Helpers/CursorPositionCalculator.ts
