+++ packages/map-editor-components/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "include": ["src", "types"],
+  "compilerOptions": {
+    "target": "es2019",
+    "module": "esnext",
+    "rootDir": "./src",
+    "strict": true,
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "experimentalDecorators": true,
+    "declaration": true,
+    "outDir": "dist"
+  }
+}
