+++ packages/map-editor-examples/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@piyoppi/pico2map-samples",
-  "version": "0.0.16",
+  "version": "0.0.17",
   "description": "",
   "main": "src/main.js",
   "directories": {
@@ -8,8 +8,8 @@
   },
   "dependencies": {
     "typescript": "^4.1.3",
-    "@piyoppi/pico2map-editor": "0.0.16",
-    "@piyoppi/pico2map-tiled": "0.0.16"
+    "@piyoppi/pico2map-ui-components": "0.0.17",
+    "@piyoppi/pico2map-tiled": "0.0.17"
   },
   "devDependencies": {
     "ts-loader": "^8.0.12",
