+++ packages/map-editor/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@piyoppi/pico2map-editor",
-  "version": "0.0.16",
+  "version": "0.0.17",
   "description": "",
   "main": "dist/main.js",
   "repository": {
@@ -8,7 +8,6 @@
     "url":"https://github.com/piyoppi/pico2map"
   },
   "dependencies": {
-    "lit-element": "^2.4.0",
     "typescript": "^4.1.3"
   },
   "devDependencies": {
@@ -18,7 +17,7 @@
     "ts-loader": "^8.0.14",
     "webpack": "^5.17.0",
     "webpack-cli": "^4.4.0",
-    "@piyoppi/pico2map-tiled": "0.0.16"
+    "@piyoppi/pico2map-tiled": "0.0.17"
   },
   "scripts": {
     "test": "jest",
