+++ packages/map-editor/src/main.ts
@@ -1,7 +1,7 @@
-export { MapChipSelectedEvent } from './Components/MapChipSelectorComponent'
-export { AutoTileSelectedEvent } from './Components/AutoTileSelectorComponent'
-export { Projects } from './Projects'
-
-import './Components/MapChipSelectorComponent'
-import './Components/MapCanvasComponent'
-import './Components/AutoTileSelectorComponent'
+export { GridImageGenerator } from './GridImageGenerator'
+export { MapCanvas } from './MapCanvas'
+export { Projects, Project } from './Projects'
+export { ColiderCanvas } from './ColiderCanvas'
+export { EditorCanvas } from './EditorCanvas'
+export { AutoTileSelector } from './AutoTileSelector'
+export { MapChipSelector } from './MapChipSelector'
